requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.11
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
fonttools==4.59.2
fpdf2==2.8.4
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==6.0.1
mercadopago==2.3.0
mysqlclient==2.2.7
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
pycparser==2.22
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
pypdf==6.0.0
python-bidi==0.6.6
python-dotenv==1.1.1
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
